// Elevate/src/app/model/checkin.model.ts
export interface CheckinVM {
  id?: number;
  assignment_id?: number;
  freelancer_id?: number;
  created_at?: string;
  created_by?: number;
  performed_at?: string;
  finished_at?: string;
  data?: any;
  is_valid?: number;
  lat?: any;
  lon?: any;
  updated_at?: string;
  updated_by?: number;
  w3w?: any;
}